Decommision DigitalOcean APIv1 and have users use the new DigitalOcean APIv2
