--- tests/integration/files/conf/cloud.providers.d/digital_ocean_v2.conf
@@ -1,5 +1,0 @@
-digitalocean-config:
-  provider: digital_ocean
-  personal_access_token: ''
-  ssh_key_file: ''
-  ssh_key_names: ''
